// services/proxy/src/models/Volo_Abp_PermissionManagement_GetPermissionListResultDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { Volo_Abp_PermissionManagement_PermissionGroupDto } from "./Volo_Abp_PermissionManagement_PermissionGroupDto";
export type Volo_Abp_PermissionManagement_GetPermissionListResultDto = {
  entityDisplayName?: string | null;
  groups?: Array<Volo_Abp_PermissionManagement_PermissionGroupDto> | null;
};