// services/proxy/src/models/Volo_Abp_PermissionManagement_PermissionGroupDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { Volo_Abp_PermissionManagement_PermissionGrantInfoDto } from "./Volo_Abp_PermissionManagement_PermissionGrantInfoDto";
export type Volo_Abp_PermissionManagement_PermissionGroupDto = {
  name?: string | null;
  displayName?: string | null;
  displayNameKey?: string | null;
  displayNameResource?: string | null;
  permissions?: Array<Volo_Abp_PermissionManagement_PermissionGrantInfoDto> | null;
};